refactor death stuff into death package
