+++ TGM/src/main/java/network/warzone/tgm/modules/death/DeathModule.java
@@ -1,4 +1,4 @@
-package network.warzone.tgm.modules;
+package network.warzone.tgm.modules.death;
 
 import lombok.NoArgsConstructor;
 import network.warzone.tgm.TGM;
