+++ TGM/src/main/java/network/warzone/tgm/modules/infection/InfectionModule.java
@@ -6,8 +6,8 @@ import network.warzone.tgm.TGM;
 import network.warzone.tgm.match.Match;
 import network.warzone.tgm.match.MatchModule;
 import network.warzone.tgm.match.MatchStatus;
-import network.warzone.tgm.modules.DeathInfo;
-import network.warzone.tgm.modules.DeathModule;
+import network.warzone.tgm.modules.death.DeathInfo;
+import network.warzone.tgm.modules.death.DeathModule;
 import network.warzone.tgm.modules.scoreboard.ScoreboardInitEvent;
 import network.warzone.tgm.modules.scoreboard.ScoreboardManagerModule;
 import network.warzone.tgm.modules.scoreboard.SimpleScoreboard;
